issue 5 loading result for a task by id implemented, also raw result could be downloaded as a file

+++ src/main/java/com/epam/blast/entity/blasttool/BlastResult.java
@@ -32,6 +32,6 @@ import java.util.List;
 @Value
 @Builder
 public class BlastResult {
-    Long size;
+    Integer size;
     List<BlastResultEntry> entries;
 }
